Check Docker image architecture before run it (#185)
* Check docker image architecture before run it

* Parse json in TS

* Add try catch and remove oclif manifest

+++ chain-cli/src/galachain-utils.ts
@@ -23,7 +23,7 @@ import { writeFile } from "node:fs/promises";
 import path from "path";
 import process from "process";
 
-import { ServicePortal } from "./consts";
+import { ExpectedImageArchitecture, ServicePortal } from "./consts";
 import { GetChaincodeDeploymentDto, PostDeployChaincodeDto } from "./dto";
 import { execSync } from "./exec-sync";
 import { parseStringOrFileKey } from "./utils";
@@ -99,6 +99,20 @@ export async function getDeploymentResponse(params: { privateKey: string; isTest
 }
 
 function getContractNames(imageTag: string): { contractName: string }[] {
+  const dockerImageInspect = execSync("docker inspect --format=json ${imageTag}");
+  let dockerJson;
+  try {
+    dockerJson = JSON.parse(dockerImageInspect);
+  } catch (e) {
+    throw new Error(`Invalid docker image inspect output: ${dockerImageInspect} - Error ${e}`);
+  }
+
+  const imageArchitecture = dockerJson[0].Os + "/" + dockerJson[0].Architecture;
+
+  if (imageArchitecture !== ExpectedImageArchitecture) {
+    throw new Error(`Unsupported architecture ${imageArchitecture}, expected ${ExpectedImageArchitecture}`);
+  }
+
   const command = `docker run --rm ${imageTag} lib/src/cli.js get-contract-names | tail -n 1`;
   let response = "<failed>";
